Revert "RM-4416"
This reverts commit 38f8fbe619cee4db68dbede650c5a394e8b2d472.

+++ i21Database/dbo/Views/vyuRKGetOpenContract.sql
@@ -9,7 +9,7 @@ FROM (
 	SELECT 
 		ot.intFutOptTransactionId
 		,dblNoOfContract =  sum(ot.dblNoOfContract)
-		,dblMatchedContract = (SELECT SUM(CONVERT(DECIMAL(18,6),mf.dblMatchQty)) FROM tblRKMatchFuturesPSDetail mf WHERE ot.intFutOptTransactionId = mf.intLFutOptTransactionId)
+		,dblMatchedContract = (SELECT SUM(CONVERT(int,mf.dblMatchQty)) FROM tblRKMatchFuturesPSDetail mf WHERE ot.intFutOptTransactionId = mf.intLFutOptTransactionId)
 	FROM tblRKFutOptTransaction ot
 	WHERE ot.strBuySell = 'Buy' AND intInstrumentTypeId = 1
 	GROUP BY intFutOptTransactionId
@@ -23,7 +23,7 @@ FROM (
 	SELECT 
 		ot.intFutOptTransactionId
 		,dblNoOfContract = sum(ot.dblNoOfContract) 
-		,dblMatchedContract = (SELECT SUM(CONVERT(DECIMAL(18,6),mf.dblMatchQty)) FROM tblRKMatchFuturesPSDetail mf WHERE ot.intFutOptTransactionId = mf.intSFutOptTransactionId)
+		,dblMatchedContract = (SELECT SUM(CONVERT(int,mf.dblMatchQty)) FROM tblRKMatchFuturesPSDetail mf WHERE ot.intFutOptTransactionId = mf.intSFutOptTransactionId)
 	FROM tblRKFutOptTransaction ot 
 	WHERE ot.strBuySell = 'Sell' AND intInstrumentTypeId = 1
 	GROUP BY intFutOptTransactionId
@@ -37,8 +37,8 @@ FROM (
 	SELECT 
 		ot.intFutOptTransactionId
 		,dblNoOfContract = sum(isnull(ot.dblNoOfContract,0))
-		,dblMatchedContract = (SELECT SUM(CONVERT(DECIMAL(18,6),mf.dblMatchQty)) FROM tblRKOptionsMatchPnS mf WHERE ot.intFutOptTransactionId=mf.intLFutOptTransactionId) 
-		,dblExpiredContract = (SELECT SUM(CONVERT(DECIMAL(18,6),ex.dblLots)) FROM tblRKOptionsPnSExpired ex WHERE ot.intFutOptTransactionId=ex.intFutOptTransactionId) 
+		,dblMatchedContract = (SELECT SUM(CONVERT(int,mf.dblMatchQty)) FROM tblRKOptionsMatchPnS mf WHERE ot.intFutOptTransactionId=mf.intLFutOptTransactionId) 
+		,dblExpiredContract = (SELECT SUM(CONVERT(int,ex.dblLots)) FROM tblRKOptionsPnSExpired ex WHERE ot.intFutOptTransactionId=ex.intFutOptTransactionId) 
 	FROM tblRKFutOptTransaction ot 
 	WHERE ot.strBuySell = 'Buy' AND intInstrumentTypeId = 2
 	GROUP BY intFutOptTransactionId
@@ -52,8 +52,8 @@ FROM (
 	SELECT 
 		ot.intFutOptTransactionId
 		,dblNoOfContract = sum(ot.dblNoOfContract) 
-		,dblMatchedContract = (SELECT SUM(CONVERT(DECIMAL(18,6),mf.dblMatchQty)) FROM tblRKOptionsMatchPnS mf WHERE ot.intFutOptTransactionId=mf.intSFutOptTransactionId)
-		,dblExpiredContract = (SELECT SUM(CONVERT(DECIMAL(18,6),ex.dblLots)) FROM tblRKOptionsPnSExpired ex WHERE ot.intFutOptTransactionId=ex.intFutOptTransactionId)
+		,dblMatchedContract = (SELECT SUM(CONVERT(int,mf.dblMatchQty)) FROM tblRKOptionsMatchPnS mf WHERE ot.intFutOptTransactionId=mf.intSFutOptTransactionId)
+		,dblExpiredContract = (SELECT SUM(CONVERT(int,ex.dblLots)) FROM tblRKOptionsPnSExpired ex WHERE ot.intFutOptTransactionId=ex.intFutOptTransactionId)
 	FROM tblRKFutOptTransaction ot 
 	WHERE ot.strBuySell = 'Sell' AND intInstrumentTypeId = 2
 	GROUP BY intFutOptTransactionId
